git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@8614 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/quantize.c
@@ -3251,8 +3251,8 @@ static int IntensityCompare(const void *x,const void *y)
 
   color_1=(PixelInfo *) x;
   color_2=(PixelInfo *) y;
-  intensity=GetPixelInfoIntensity(color_1)-(ssize_t)
-    GetPixelInfoIntensity(color_2);
+  intensity=(ssize_t) (GetPixelInfoIntensity(color_1)-(ssize_t)
+    GetPixelInfoIntensity(color_2));
   return((int) intensity);
 }
